% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/matrixInput.R
\name{matrixInput}
\alias{matrixInput}
\title{Create a matrix input field}
\usage{
matrixInput(inputId, value = matrix("", 1, 1), inputClass = "",
  rows = list(), cols = list(), class = "character", paste = FALSE,
  copy = FALSE)
}
\arguments{
\item{inputId}{The input slot that will be used to access the value}

\item{value}{Inital value. Should be a matrix}

\item{inputClass}{class of the matrix input html element}

\item{rows}{list of options to configure rows}

\item{cols}{list of options to configure cols}

\item{class}{Matrix will be coerced to a matrix of this class. `character` and `numeric`
are supported}

\item{paste}{enable paste functionality}

\item{copy}{enable copy functionality}
}
\description{
Create a matrix input field
}
