% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/startApplication.R
\name{startApplication}
\alias{startApplication}
\title{Start Application}
\usage{
startApplication(app = "app", port = 4242)
}
\arguments{
\item{app}{name of the folder in `inst`}

\item{port}{port of web application}
}
\description{
This function starts an example app from the folder `inst`.
}
\examples{
startApplication("appCopy")

}
