% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/matrixInput.R
\name{updateMatrixInput}
\alias{updateMatrixInput}
\title{Update matrix input}
\usage{
updateMatrixInput(session, inputId, value)
}
\arguments{
\item{session}{shiny session}

\item{inputId}{id of matrix input}

\item{value}{new value for matrix}
}
\description{
This function updates the matrix input from R created with `matrixInput`.
It works like the other updateXXXInput functions in shiny.
}
\examples{
\dontrun{
updateMatrixInput(session, "myMatrix", diag(4))
}

}
