% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/matrixInput.R
\name{matrixInput}
\alias{matrixInput}
\title{Create a matrix input field}
\usage{
matrixInput(inputId, rows = 4, cols = 4, type = "character")
}
\arguments{
\item{inputId}{The input slot that will be used to access the value}

\item{rows}{Number of rows}

\item{cols}{Number of cols}

\item{type}{class of matrix}
}
\description{
Create a matrix input field
}
