% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/matrixInput.R
\name{updateMatrixInput}
\alias{updateMatrixInput}
\title{Update matrix input}
\usage{
updateMatrixInput(session, inputId, value = NULL)
}
\arguments{
\item{session}{shiny session}

\item{inputId}{id of matrix input}

\item{value}{new value for matrix}
}
\description{
Update matrix input
}
